Remove unnecessary dependencies `syn` and `quote`
`syn` had some limitations in what it could parse (easily). As this is a
simple crate and does not need `syn`'s more featureful parsing, it has
been replaced so that the macro can be used more flexibly. Likewise, the
`quote` crate was removed as the restructuring to remove `syn` made it
unnecessary .